+++ test/bones_api_mysql_test.dart
@@ -1,10 +1,8 @@
 @TestOn('vm')
 @Tags(['docker', 'mysql', 'slow'])
 @Timeout(Duration(seconds: 180))
-import 'package:bones_api/bones_api.dart';
 import 'package:bones_api/bones_api_adapter_mysql.dart';
-import 'package:docker_commander/docker_commander.dart';
-import 'package:docker_commander/docker_commander_vm.dart';
+import 'package:bones_api/bones_api_test_mysql.dart';
 import 'package:test/test.dart';
 
 import 'bones_api_test_adapter.dart';
@@ -13,63 +11,20 @@ final dbUser = 'myuser';
 final dbPass = 'mypass';
 final dbName = 'mydb';
 
-class MySQLTestContainer extends DBTestContainerDocker {
-  @override
-  String get name => 'mysql';
-
-  late final MySQLContainerConfig containerConfig;
-  late MySQLContainer container;
-
-  @override
-  Future<bool> start(int dbPort) async {
-    containerConfig = MySQLContainerConfig(
-      dbUser: dbUser,
-      dbPassword: dbPass,
-      dbName: dbName,
-      hostPort: dbPort,
-      forceNativePasswordAuthentication: true,
-    );
-
-    container = await containerConfig.run(containerHandler!,
-        name: 'dc_test_mysql_$dbPort', cleanContainer: true);
-    return true;
-  }
-
-  @override
-  Future<bool> waitReady() => container.waitReady();
-
-  static const String configFile = '/etc/mysql/my.cnf';
-  static const String configDirectory = '/etc/mysql/conf.d';
-
-  @override
-  Future<String?> prepare() async {
-    var out = await container.execCat(configFile);
-    var out2 = await container
-        .execAndWaitStdoutAsString('ls', ['-al', configDirectory]);
-    return '\n*** $configFile:\n$out\n*** configDirectory:\n$out2';
-  }
-
-  @override
-  Future<String?> finalize() async {
-    var out1 = await container.mysqlCMD('SHOW TABLES');
-    var out2 = await container.mysqlCMD('select * from `user`;');
-    return '$out1\n$out2';
-  }
-
-  @override
-  Future<bool> stop() => container.stop(timeout: Duration(seconds: 30));
-
-  @override
-  Future<String?> runSQL(String sqlInline) => container.runSQL(sqlInline);
-
-  @override
-  Future<String> listTables() async {
-    var out = await container.mysqlCMD('SHOW TABLES');
-    return out ?? '';
-  }
-
-  @override
-  String get stdout => container.stdout?.asString ?? '';
+class MySQLTestConfig extends APITestConfigDockerMySQL {
+  MySQLTestConfig()
+      : super({
+          'db': {
+            'mysql': {
+              'username': dbUser,
+              'password': dbPass,
+              'database': dbName,
+              'port': -3306,
+            }
+          }
+        },
+            containerNamePrefix: 'bones_api_test_mysql',
+            forceNativePasswordAuthentication: true);
 }
 
 Future<void> main() async {
@@ -79,8 +34,7 @@ Future<void> main() async {
 
 Future<bool> _runTest(bool useReflection) => runAdapterTests(
       'MySQL',
-      MySQLTestContainer(),
-      3306,
+      MySQLTestConfig(),
       (provider, dbPort) => MySQLAdapter(
         dbName,
         dbUser,